backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["e.truhin <65777835+TeaArx@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
python-dotenv = "^1.0.1"
drf-spectacular = "^0.27.2"
markdown = "^3.7"
django-filter = "^24.3"
psycopg2-binary = "^2.9.10"
djoser = "^2.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
